feat: adding revision chore: unit tests
